// src/types/interfaces/CurrentConditionData.d.ts
export interface CurrentConditionData {
// ...
    Wind: {
        Direction: {
            Degrees: number; //Wind direction in Azimuth degrees (e.g. 180 degrees is a wind coming from the south). May be NULL.
            Localized: string; //Direction abbreviated in the language specified by language code in URL.
            English: Directions | string; //Direction abbreviated in English.
        };
        Speed: {
            Metric: Values<"km/h", 7>;
            Imperial: Values<"mi/h", 9>;
        }; //Wind Speed. Contains Metric and Imperial Values.
    };
// ...
}